Fix SUB_KTEMP_WHITE algorithm: correct white drive and RGB subtraction
The previous implementation used the raw limiting channel value (uScale)
as the white LED drive, which undershot by a factor of sumW/wlim. The
correct drive under the 1/3-efficiency model is fRatio*sumW, capped at
255 with fActualRatio back-calculated for the RGB subtraction.

Also fix div-by-zero on low color temperatures (zero white channel
components), replace non-standard u_int16_t/u_int8_t with uint16_t/
uint8_t, fix invalid default whiteTemp=0 -> 5000, fix _customWhite-
Temperature type int -> uint16_t, and improve readability of
ColorRgb::white().

+++ libsrc/leddevice/dev_net/LedDeviceUdpE131.h
@@ -161,7 +161,7 @@ private:
 	RGBW::WhiteAlgorithm _whiteAlgorithm;
 	ColorRgbw _temp_rgbw;
 	uint16_t _dmxChannelCount;
-	int _customWhiteTemperature = 5000;
+	uint16_t _customWhiteTemperature = 5000;
 };
 
 #endif // LEDEVICEUDPE131_H
